Update Linux installer to 10.1
First, the Linux installer is kind of broken, we need to add a new
option in order for all of the libraries to be installed. This is
discussed on the nvidia forums [here](https://devtalk.nvidia.com/default/topic/1047903/linux/new-cuda-10-1-run-installer-completely-broken/).

Second, the URL for the installer has changed. The file now has a 'run'
extension.